requirements.txt
aiogram==3.2.0
asyncpg==0.28.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pydantic==2.4.2
loguru==0.7.2
pytz==2023.3
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
apscheduler
PyJWT
